08/120/2017 17:34; Final Commit - App completed and functional.
Note: This commit has been made from a different mac as my laptop crashed and caused scroll.png and title2.png to become corrupted. These files had to be deleted and added again. Furthermore, Xcode no longer runs on my laptop as of 4:45 pm and says it is being updated. I have not found a way to stop the update and will have to investigate the error further.